requirements.txt
Pillow>=9.0.0
matplotlib>=3.5.0
numpy>=1.21.0
pillow-heif>=0.16.0
